Fixed jelly files that configure the steps.
