+++ src/main/java/io/jenkins/plugins/analysis/core/history/ResultSelector.java
@@ -2,12 +2,12 @@ package io.jenkins.plugins.analysis.core.history;
 
 import java.util.Optional;
 
-import io.jenkins.plugins.analysis.core.steps.PipelineResultAction;
+import io.jenkins.plugins.analysis.core.steps.ResultAction;
 
 import hudson.model.Run;
 
 /**
- * Selects a {@link PipelineResultAction} from all registered actions in a given job.
+ * Selects a {@link ResultAction} from all registered actions in a given job.
  *
  * @author Ullrich Hafner
  */
@@ -18,5 +18,5 @@ public interface ResultSelector {
      * @param run the run
      * @return the result action, if there is one attached to the job
      */
-    Optional<PipelineResultAction> get(Run<?, ?> run);
+    Optional<ResultAction> get(Run<?, ?> run);
 }
