+++ src/main/java/io/jenkins/plugins/analysis/core/steps/StaticAnalysis.java
@@ -3,7 +3,6 @@ package io.jenkins.plugins.analysis.core.steps;
 import java.io.File;
 import java.lang.reflect.InvocationTargetException;
 import java.util.Collection;
-import java.util.Collections;
 
 import hudson.Extension;
 import hudson.plugins.analysis.util.model.FileAnnotation;
@@ -21,6 +20,6 @@ public class StaticAnalysis extends StaticAnalysisTool {
 
     @Override
     public Collection<FileAnnotation> parse(final File file, final String moduleName) throws InvocationTargetException {
-        return Collections.emptyList();
+        throw new UnsupportedOperationException("Static analysis generic tool does not support parsing");
     }
 }
