+++ src/main/resources/io/jenkins/plugins/analysis/core/steps/JobAction/floatingBox.jelly
@@ -1,6 +1,6 @@
 <?jelly escape-by-default='true'?>
-<j:jelly xmlns:j="jelly:core" xmlns:st="jelly:stapler"
->
+<j:jelly xmlns:j="jelly:core" xmlns:st="jelly:stapler" xmlns:d="jelly:define" xmlns:l="/lib/layout"
+  xmlns:t="/lib/hudson" xmlns:f="/lib/form" xmlns:i="jelly:fmt" xmlns:local="local">
   <j:if test="${from.isTrendVisible(request)}">
     <div align="right">
       <div class="test-trend-caption">
@@ -9,12 +9,11 @@
       <div>
         <!-- Note that the parameter ?url=${from.trendGraphId} is not used by the graph: it
              has been added to provide different URLs so that the images are not cached by the browser. -->
-        <img src="${from.urlName}/trendGraph/png?url=${from.trendGraphId}"
-             lazymap="${from.urlName}/trendGraph/map?url=${from.trendGraphId}"/>
+        <img src="${from.urlName}/trendGraph/png?url=${from.trendGraphId}" lazymap="${from.urlName}/trendGraph/map?url=${from.trendGraphId}" />
       </div>
       <div>
         <a href="${from.urlName}/trendDetails">${%Enlarge}</a>
-        <st:nbsp/>
+        <st:nbsp />
         <a href="${from.urlName}/configure">${%Configure}</a>
       </div>
     </div>
