--- src/main/resources/io/jenkins/plugins/analysis/core/steps/ParseWarningsStep/config.jelly
@@ -1,10 +1,0 @@
-<?jelly escape-by-default='true'?>
-<j:jelly xmlns:j="jelly:core" xmlns:st="jelly:stapler" xmlns:d="jelly:define"
-         xmlns:l="/lib/layout" xmlns:t="/lib/hudson" xmlns:f="/lib/form" xmlns:u="/util">
-
-  <u:detectModules/>
-  <u:defaultEncoding id="warnings"/>
-
-  <f:dropdownDescriptorSelector field="parser" title="Hallo" descriptors="${descriptor.availableParsers}"/>
-
-</j:jelly>
