--- src/main/resources/io/jenkins/plugins/analysis/core/steps/PipelineResultAction/summary.jelly
@@ -1,5 +1,0 @@
-<?jelly escape-by-default='true'?>
-<j:jelly xmlns:j="jelly:core" xmlns:st="jelly:stapler" xmlns:d="jelly:define" xmlns:l="/lib/layout"
-  xmlns:t="/lib/hudson" xmlns:f="/lib/form" xmlns:i="jelly:fmt" xmlns:u="/util">
-  <u:summary />
-</j:jelly>
