--- src/main/resources/io/jenkins/plugins/analysis/core/steps/PublishWarningsStep/config.properties
@@ -1,11 +1,0 @@
-title.onlyUseStableBuildsAsReference=Only use stable builds as reference
-description.onlyUseStableBuildsAsReference=Use the last stable build as the reference to compute the number of new \
-                   warnings against. This allows to ignore interim unstable builds for which the number of warnings \
-                   decreased. Note that the last stable build is evaluated only by inspecting the unit test failures. \
-                   The static analysis results are not considered.
-title.usePreviousBuildAsReference=Use previous build as reference
-description.usePreviousBuildAsReference=If set the number of new warnings will always be computed based on the \
-                   previous build, even if that build is unstable (due to a violated warning threshold). \
-                   Otherwise the last build that did not violate any \
-                   given threshold will be used as reference. It is recommended to uncheck this option if \
-                   the plug-in should ensure that all new warnings will be finally fixed in subsequent builds.
