--- src/main/resources/io/jenkins/plugins/analysis/core/steps/PublishWarningsStep/config_de.properties
@@ -1,9 +1,0 @@
-title.onlyUseStableBuildsAsReference=Nur stabile Builds als Referenz nutzen
-description.onlyUseStableBuildsAsReference=Die Anzahl der neuen Warnungen wird auf Basis des letzten stabilen Builds \
-         bestimmt (Achtung: die Stabilität bezieht sich nicht auf die statische Analyse sondern nur auf die Ergebnisse \
-         der Unit Tests). Dadurch lässt sich ein instabiler Build rückgängig machen, ohne dass die Berechnung der Warnungen \
-         beeinflusst wird.
-title.usePreviousBuildAsReference=Letzten Build als Referenz nutzen
-description.usePreviousBuildAsReference=Falls aktiviert, wird die Anzahl der neuen Warnungen auf Basis des letzten Builds \
-           bestimmt. Normalerweise wird die Anzahl der neuen Warnungen durch einen Vergleich mit dem Referenzbuild ermittelt,\
-           d.h. mit dem Build, bei dem die neuen Warnungen das erste Mal erschienen sind.
